+++ src/schema/types.ts
@@ -129,6 +129,10 @@ type AsyncSetter<T, Input, Output, CtxOptions extends ObjectType> = (
 
 type NotAllowedError = string | InputFieldError;
 
+type SetterWithSummary<T, Input, Output, CtxOptions extends ObjectType> = (
+  summary: MutableSummary<Input, Output, CtxOptions> & {},
+) => TypeOf<T>;
+
 type Setter<T, Input, Output, CtxOptions extends ObjectType> = (
   context: MutableContext<Input, Output, CtxOptions>,
 ) => TypeOf<T>;
@@ -166,26 +170,46 @@ type VirtualResolver<
   summary: MutableSummary<Input, Output, CtxOptions>,
 ) => TypeOf<Input[K]> | Promise<TypeOf<Input[K]>>;
 
-type PostValidator<Input, Output, Aliases, CtxOptions extends ObjectType> = (
+type PostValidator<
+  InputKeys extends KeyOf<Input>,
+  Input,
+  Output,
+  Aliases,
+  CtxOptions extends ObjectType,
+> = (
   summary: MutableSummary<Input, Output, CtxOptions>,
-  propertiesProvided: KeyOf<Input>[],
+  propertiesProvided: InputKeys[],
 ) =>
   | undefined
   | ResponseErrorObject<Input, Aliases>
-  | Promise<undefined | ResponseErrorObject<Input, Aliases>>;
+  | PostValidatorSanitizedResponse<InputKeys, Input, Output>
+  | Promise<
+      | undefined
+      | ResponseErrorObject<Input, Aliases>
+      | PostValidatorSanitizedResponse<InputKeys, Input, Output>
+    >;
+
+type PostValidatorSanitizedResponse<K extends KeyOf<Input>, Input, Output> = {
+  [Key in K]?: {
+    validated: TypeOf<Key extends KeyOf<Output> ? Output[Key] : Input[Key]>;
+  };
+};
 
 type PostValidationConfig<
+  K extends KeyOf<Input>,
   Input,
   Output,
   Aliases,
   CtxOptions extends ObjectType,
 > = {
-  properties: ArrayOfMinSizeTwo<KeyOf<Input>>;
+  properties: ArrayOfMinSizeTwo<K>;
   validator:
-    | PostValidator<Input, Output, Aliases, CtxOptions>
+    | PostValidator<K, Input, Output, Aliases, CtxOptions>
     | ArrayOfMinSizeOne<
-        | PostValidator<Input, Output, Aliases, CtxOptions>
-        | ArrayOfMinSizeOne<PostValidator<Input, Output, Aliases, CtxOptions>>
+        | PostValidator<K, Input, Output, Aliases, CtxOptions>
+        | ArrayOfMinSizeOne<
+            PostValidator<K, Input, Output, Aliases, CtxOptions>
+          >
       >;
 };
 
@@ -306,6 +330,7 @@ namespace NS {
       constant?: unknown;
       default?: unknown;
       dependsOn?: KeyOf<Input> | KeyOf<Input>[];
+      ignore?: SetterWithSummary<boolean, Input, Output, {}>;
       readonly?: boolean | 'lax';
       resolver?: Function;
       required?: boolean | RequiredHandler<Input, Output, {}>;
@@ -405,19 +430,22 @@ namespace NS {
     Output,
     CtxOptions extends ObjectType,
   > = XOR<
-    {
-      shouldInit?: Setter<boolean, Input, Output, CtxOptions>;
-      shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
-    },
+    { ignore?: SetterWithSummary<boolean, Input, Output, CtxOptions> },
     XOR<
       {
-        shouldInit?: false | Setter<boolean, Input, Output, CtxOptions>;
+        shouldInit?: Setter<boolean, Input, Output, CtxOptions>;
         shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
       },
-      {
-        shouldInit?: Setter<boolean, Input, Output, CtxOptions>;
-        shouldUpdate?: false | Setter<boolean, Input, Output, CtxOptions>;
-      }
+      XOR<
+        {
+          shouldInit?: false | Setter<boolean, Input, Output, CtxOptions>;
+          shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
+        },
+        {
+          shouldInit?: Setter<boolean, Input, Output, CtxOptions>;
+          shouldUpdate?: false | Setter<boolean, Input, Output, CtxOptions>;
+        }
+      >
     >
   >;
 
@@ -449,14 +477,19 @@ namespace NS {
       | TypeOf<Output[K]>
       | AsyncSetter<Output[K], Input, Output, CtxOptions>;
     readonly: 'lax';
-    shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
     validator:
       | Validator<K, Input, Output, CtxOptions>
       | [
           Validator<K, Input, Output, CtxOptions>,
           SecondaryValidator<Output[K], Input, Output, CtxOptions>,
         ];
-  };
+  } & XOR<
+      { ignore?: SetterWithSummary<boolean, Input, Output, CtxOptions> },
+      {
+        shouldInit?: Setter<boolean, Input, Output, CtxOptions>;
+        shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
+      }
+    >;
 
   type ReadonlyNoInit<
     K extends keyof (Output | Input),
@@ -468,15 +501,19 @@ namespace NS {
       | TypeOf<Output[K]>
       | AsyncSetter<Output[K], Input, Output, CtxOptions>;
     readonly: true;
-    shouldInit: false | Setter<boolean, Input, Output, CtxOptions>;
-    shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
     validator?:
       | Validator<K, Input, Output, CtxOptions>
       | [
           Validator<K, Input, Output, CtxOptions>,
           SecondaryValidator<Output[K], Input, Output, CtxOptions>,
         ];
-  };
+  } & XOR<
+      { ignore?: SetterWithSummary<boolean, Input, Output, CtxOptions> },
+      {
+        shouldInit?: false | Setter<boolean, Input, Output, CtxOptions>;
+        shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
+      }
+    >;
 
   type RequiredReadonly<
     K extends keyof (Output | Input),
@@ -542,15 +579,19 @@ namespace NS {
       | AsyncSetter<Output[K], Input, Output, CtxOptions>;
     required: RequiredHandler<Input, Output, CtxOptions>;
     readonly?: true;
-    shouldInit?: Setter<boolean, Input, Output, CtxOptions>;
-    shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
     validator:
       | Validator<K, Input, Output, CtxOptions>
       | [
           Validator<K, Input, Output, CtxOptions>,
           SecondaryValidator<Output[K], Input, Output, CtxOptions>,
         ];
-  };
+  } & XOR<
+      { ignore?: SetterWithSummary<boolean, Input, Output, CtxOptions> },
+      {
+        shouldInit?: Setter<boolean, Input, Output, CtxOptions>;
+        shouldUpdate?: Setter<boolean, Input, Output, CtxOptions>;
+      }
+    >;
 
   type Virtual<
     K extends keyof Input,
@@ -595,9 +636,9 @@ namespace NS {
       | SuccessHandler<Input, Output, CtxOptions>
       | ArrayOfMinSizeOne<SuccessHandler<Input, Output, CtxOptions>>;
     postValidate?:
-      | PostValidationConfig<Input, Output, object, CtxOptions>
+      | PostValidationConfig<KeyOf<Input>, Input, Output, object, CtxOptions>
       | ArrayOfMinSizeOne<
-          PostValidationConfig<Input, Output, object, CtxOptions>
+          PostValidationConfig<KeyOf<Input>, Input, Output, object, CtxOptions>
         >;
     setMissingDefaultsOnUpdate?: boolean;
     shouldUpdate?: boolean | AsyncShouldUpdate<Input, Output, CtxOptions>;
@@ -620,9 +661,9 @@ namespace NS {
       | ArrayOfMinSizeOne<DeleteHandler<Output, CtxOptions>>;
     onSuccess?: OnSuccessConfig<Input, Output, CtxOptions>;
     postValidate?:
-      | PostValidationConfig<Input, Output, Aliases, CtxOptions>
+      | PostValidationConfig<KeyOf<Input>, Input, Output, Aliases, CtxOptions>
       | ArrayOfMinSizeOne<
-          PostValidationConfig<Input, Output, Aliases, CtxOptions>
+          PostValidationConfig<KeyOf<Input>, Input, Output, Aliases, CtxOptions>
         >;
     setMissingDefaultsOnUpdate?: boolean;
     shouldUpdate?: boolean | AsyncShouldUpdate<Input, Output, CtxOptions>;
@@ -729,6 +770,7 @@ const DEFINITION_RULES = [
   'constant',
   'default',
   'dependsOn',
+  'ignore',
   'onDelete',
   'onFailure',
   'onSuccess',
@@ -759,6 +801,7 @@ const CONSTANT_RULES = ['constant', 'onDelete', 'onSuccess', 'value'];
 const VIRTUAL_RULES = [
   'alias',
   'allow',
+  'ignore',
   'sanitizer',
   'onFailure',
   'onSuccess',